table and pagination

+++ Frontend/src/app/Components/students/students.component.ts
@@ -1,26 +1,35 @@
-import {
-  Component,
-  OnInit,
-  ViewChild,
-} from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { ModalComponent } from '../modal/modal.component';
 import { StudentsService } from 'src/app/Services/students.service';
 import { Student } from 'src/models/students.model';
 import { CommonModule } from '@angular/common';
+import {
+  NgbPaginationModule,
+  NgbTypeaheadModule,
+} from '@ng-bootstrap/ng-bootstrap';
+import { FormsModule } from '@angular/forms';
 
 @Component({
   standalone: true,
   selector: 'app-students',
   templateUrl: './students.component.html',
   styleUrls: ['./students.component.css'],
-  imports: [ModalComponent, CommonModule],
+  imports: [
+    ModalComponent,
+    CommonModule,
+    NgbTypeaheadModule,
+    NgbPaginationModule,
+    FormsModule,
+  ],
 })
 export class StudentsComponent implements OnInit {
   students: Student[] | undefined;
   student: Student | undefined;
-  dateNow = new Date().getTime();
-  test: any;
   @ViewChild(ModalComponent) modal: ModalComponent | undefined;
+  page = 1;
+  pageSize = 4;
+  collectionSize = 0;
+  paginatedStudents: Student[] | undefined;
   constructor(private studentsService: StudentsService) {}
   ngOnInit(): void {
     this.getAllStudents();
@@ -39,6 +48,10 @@ export class StudentsComponent implements OnInit {
     this.studentsService.GetAllStudents().subscribe({
       next: (data: any) => {
         this.students = data;
+        if (this.students) {
+          this.collectionSize = this.students.length;
+        }
+        this.refreshStudents();
       },
       error: (err) => {
         console.log(err);
@@ -49,13 +62,27 @@ export class StudentsComponent implements OnInit {
     });
   }
   deleteStudent(id: string) {
-    this.studentsService.RemoveStudent(id).subscribe();
-    this.getAllStudents();
+    const confirmDelete = confirm(
+      'Are you sure you want to delete this student?'
+    );
+    if (confirmDelete) {
+      this.studentsService.RemoveStudent(id).subscribe();
+      this.getAllStudents();
+    }
   }
   getAge(bdate: any) {
+    const dateNow = new Date().getTime();
     const birthday = new Date(bdate).getTime();
-    const diff = this.dateNow - birthday;
+    const diff = dateNow - birthday;
     const age = Math.floor(diff / 31556952000);
     return age;
+  }
+  refreshStudents() {
+    this.paginatedStudents = this.students
+      ?.map((student: any, i: any) => ({ id: i + 1, ...student }))
+      .slice(
+        (this.page - 1) * this.pageSize,
+        (this.page - 1) * this.pageSize + this.pageSize
+      );
   }
 }
